+++ plugins/destination/clickhouse/queries/insert.go
@@ -3,12 +3,11 @@ package queries
 import (
 	"strings"
 
-	"github.com/apache/arrow/go/v13/arrow"
 	"github.com/cloudquery/cloudquery/plugins/destination/clickhouse/util"
-	"github.com/cloudquery/plugin-sdk/v2/schema"
+	"github.com/cloudquery/plugin-sdk/v3/schema"
 )
 
-func Insert(sc *arrow.Schema) string {
-	return `INSERT INTO ` + util.SanitizeID(schema.TableName(sc)) +
-		`(` + strings.Join(util.Sanitized(ColumnNames(sc.Fields())...), `, `) + `)`
+func Insert(table *schema.Table) string {
+	return `INSERT INTO ` + util.SanitizeID(table.Name) +
+		`(` + strings.Join(util.Sanitized(table.Columns.Names()...), `, `) + `)`
 }