+++ plugins/source/aws/client/services.go
@@ -5,6 +5,7 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/accessanalyzer"
 	"github.com/aws/aws-sdk-go-v2/service/account"
 	"github.com/aws/aws-sdk-go-v2/service/acm"
+	"github.com/aws/aws-sdk-go-v2/service/acmpca"
 	"github.com/aws/aws-sdk-go-v2/service/amp"
 	"github.com/aws/aws-sdk-go-v2/service/amplify"
 	"github.com/aws/aws-sdk-go-v2/service/apigateway"
@@ -117,6 +118,7 @@ func initServices(c aws.Config, regions []string) Services {
 		Accessanalyzer:            accessanalyzer.NewFromConfig(awsCfg),
 		Account:                   account.NewFromConfig(awsCfg),
 		Acm:                       acm.NewFromConfig(awsCfg),
+		Acmpca:                    acmpca.NewFromConfig(awsCfg),
 		Amp:                       amp.NewFromConfig(awsCfg),
 		Amplify:                   amplify.NewFromConfig(awsCfg),
 		Apigateway:                apigateway.NewFromConfig(awsCfg),
@@ -227,6 +229,7 @@ type Services struct {
 	Accessanalyzer            services.AccessanalyzerClient
 	Account                   services.AccountClient
 	Acm                       services.AcmClient
+	Acmpca                    services.AcmpcaClient
 	Amp                       services.AmpClient
 	Amplify                   services.AmplifyClient
 	Apigateway                services.ApigatewayClient